Response code changed
I think that is better to use 405 as response code in this case. That's because the reason for this error is that the method is not found/allowed. 
It will be good to have separate cases if method is not declared in the controller and when the HTTP method does not match the expected one.
I also purpose to use exclamation mark in the if instead of ```method_exists($this, $controller_method) === FALSE```. It is shorter and since (method_exits())[http://php.net/manual/en/function.method-exists.php] returns boolean, I do not see a problem to do it ;).